// src/app/models/lugar.ts
export interface Lugares {
   id_lugar: string ;
   nombre_lugar?: string;
   latitud:string;
   longitud: string;
   descripcion?: string;
   calle?: string;
   colonia?: string;
   codigo_postal?: string;
   no_exterior?: string;
   no_interior?: string;
   telefono?: string;
   id_categoria?: string;
}